feat(play): extract the v0play hash protocol as @vuetify/play
One encoder, decoder, and theme sanitizer for docs, v0play, and later
genesis/builder. ThemeAdapter.SAFE_IDENT and UNSAFE_CSS are public; CSS
comments are rejected in theme color values. Play-hash ids stay
letter-prefixed and skip UNSAFE_KEYS.
