+++ apps/docs/package.json
@@ -54,8 +54,8 @@
     "@vue/repl": "catalog:",
     "@vuetify/auth": "catalog:",
     "@vuetify/paper": "workspace:*",
+    "@vuetify/play": "workspace:*",
     "@vuetify/v0": "workspace:*",
-    "fflate": "catalog:",
     "markdown-it-container": "catalog:",
     "marked": "catalog:",
     "marked-emoji": "catalog:",
