+++ apps/playground/src/composables/usePlayground.ts
@@ -1,114 +1,40 @@
+import {
+  buildPlaygroundFiles,
+  encodePlaygroundHash,
+  isFileRecord,
+  parsePlaygroundPayload,
+  unzipPlaygroundHash,
+} from '@vuetify/play'
+
 // Framework
-import { isArray, isBoolean, isObject, isString, isUndefined } from '@vuetify/v0'
+import { isArray, isObject, isString } from '@vuetify/v0'
 
-// Utilities
-import { toPascal } from '@/utilities/strings'
+export {
+  buildPlaygroundFiles,
+  detectEntryFile,
+  encodePlaygroundHash,
+  generateAppWrapper,
+  isFileRecord,
+  loadFflate,
+  parsePlaygroundPayload,
+  sanitizePlaygroundThemes,
+  toPlaygroundThemes,
+  unzipPlaygroundHash,
+} from '@vuetify/play'
+export type {
+  PlaygroundFile,
+  PlaygroundHashData,
+  PlaygroundHashSettings,
+  PlaygroundThemeDefinition,
+  UsePlaygroundOptions,
+} from '@vuetify/play'
 
-export interface PlaygroundFile {
-  name: string
-  code: string
-}
-
-// eslint-disable-next-line @typescript-eslint/consistent-type-imports
-let _fflate: typeof import('fflate') | undefined
-
-export async function loadFflate () {
-  if (!_fflate) _fflate = await import('fflate')
-  return _fflate
-}
-
-async function utoa (data: string): Promise<string> {
-  const { strToU8, strFromU8, zlibSync } = await loadFflate()
-  const buffer = strToU8(data)
-  const zipped = zlibSync(buffer, { level: 9 })
-  const binary = strFromU8(zipped, true)
-  return btoa(binary)
-}
-
-async function atou (base64: string): Promise<string> {
-  const { strToU8, strFromU8, unzlibSync } = await loadFflate()
-  const binary = atob(base64)
-  const buffer = strToU8(binary, true)
-  const unzipped = unzlibSync(buffer)
-  return strFromU8(unzipped)
-}
-
-/**
- * Detect which file is the entry point for a multi-file example.
- */
-export function detectEntryFile (files: PlaygroundFile[]): PlaygroundFile | undefined {
-  const vueFiles = files.filter(f => f.name.endsWith('.vue'))
-
-  const entryNames = ['index.vue', 'App.vue', 'example.vue', 'main.vue']
-  for (const name of entryNames) {
-    const found = vueFiles.find(f => f.name.toLowerCase() === name.toLowerCase())
-    if (found) return found
-  }
-
-  for (const file of vueFiles) {
-    const importsOthers = vueFiles.some(other =>
-      other !== file && file.code.includes(`./${other.name.replace(/\.\w+$/, '')}`),
-    )
-    if (importsOthers) return file
-  }
-
-  return vueFiles.at(-1)
-}
-
-/**
- * Generate an App.vue wrapper that imports and renders the entry component.
- */
-export function generateAppWrapper (entryPath: string): string {
-  const baseName = entryPath.split('/').pop()!.replace(/\.vue$/, '')
-  const pascalName = toPascal(baseName)
-  return [
-    '<' + `script setup lang="ts">`,
-    `  import ${pascalName} from './${entryPath}'`,
-    '</' + 'script>',
-    '',
-    '<template>',
-    '  <div class="p-4">',
-    `    <${pascalName} />`,
-    '  </div>',
-    '</template>',
-    '',
-  ].join('\n')
-}
-
-/**
- * Build the src/-prefixed file record that loadExample expects.
- * When dir is provided, files are nested: src/{dir}/{name}
- */
-export function buildPlaygroundFiles (inputFiles: PlaygroundFile[], dir?: string): Record<string, string> {
-  const files: Record<string, string> = {}
-  const prefix = dir ? `src/${dir}` : 'src'
-
-  for (const file of inputFiles) {
-    const path = file.name.startsWith('src/') ? file.name : `${prefix}/${file.name}`
-    files[path] = file.code
-  }
-
-  const hasAppVue = inputFiles.some(f => f.name.toLowerCase() === 'app.vue')
-  if (!hasAppVue) {
-    const entryFile = detectEntryFile(inputFiles)
-    if (entryFile) {
-      const entryPath = dir ? `${dir}/${entryFile.name}` : entryFile.name
-      files['src/App.vue'] = generateAppWrapper(entryPath)
-    }
-  }
-
-  return files
-}
-
-export interface UsePlaygroundOptions {
-  dir?: string
-  imports?: Record<string, string>
-  settings?: PlaygroundHashSettings
-  /** Selected sandbox theme id. */
-  theme?: string
-  /** Theme records merged into the sandbox `createThemePlugin`. */
-  themes?: Record<string, PlaygroundThemeDefinition>
-}
+// Types
+import type {
+  PlaygroundFile,
+  PlaygroundHashData,
+  UsePlaygroundOptions,
+} from '@vuetify/play'
 
 /**
  * Build a same-origin v0play hash URL for the given files.
@@ -147,10 +73,6 @@ export async function usePlayground (
   if (options.themes && Object.keys(options.themes).length > 0) data.themes = options.themes
   const hash = await encodePlaygroundHash(data)
   return `/#${hash}`
-}
-
-export function isFileRecord (v: unknown): v is Record<string, string> {
-  return isObject(v) && Object.values(v).every(x => isString(x))
 }
 
 /**
@@ -242,160 +164,17 @@ export function parseVuetifyPlayTuple (parsed: unknown[]): { files: Record<strin
   }
 }
 
-export interface PlaygroundThemeDefinition {
-  dark: boolean
-  colors: Record<string, string>
-}
-
-export interface PlaygroundHashSettings {
-  vue?: string
-  v0?: string
-  vuetify?: string
-  vuetifyNightly?: boolean
-  preset?: string
-  addons?: string
-}
-
-export interface PlaygroundHashData {
-  files: Record<string, string>
-  active?: string
-  imports?: Record<string, string>
-  settings?: PlaygroundHashSettings
-  /** Selected sandbox theme id. */
-  theme?: string
-  /** Theme records merged into the sandbox `createThemePlugin`. */
-  themes?: Record<string, PlaygroundThemeDefinition>
-}
-
-/** Theme ids written into generated `main.ts` / hash JSON. */
-export const SAFE_THEME_ID = /^[a-zA-Z][\w-]*$/
-
-const SAFE_COLOR_KEY = /^[a-zA-Z0-9_-]+$/
-const UNSAFE_CSS = /url\s*\(|src\s*\(|image\s*\(|image-set\s*\(|cross-fade\s*\(|@import|expression\s*\(|[;{}<>\\]|\/\*/i
-const UNSAFE_OBJECT_KEYS = new Set(['__proto__', 'constructor', 'prototype'])
-const MAX_THEME_ID = 64
-const MAX_THEMES = 32
-const MAX_COLORS = 64
-const MAX_COLOR_VALUE = 128
-
-function sanitizeThemeId (id: string): string | undefined {
-  return id.length <= MAX_THEME_ID && SAFE_THEME_ID.test(id) && !UNSAFE_OBJECT_KEYS.has(id)
-    ? id
-    : undefined
-}
-
-function sanitizeColors (colors: Record<string, unknown>): Record<string, string> {
-  const out: Record<string, string> = {}
-  for (const [key, value] of Object.entries(colors)) {
-    if (Object.keys(out).length >= MAX_COLORS) break
-    if (
-      key.length > MAX_THEME_ID
-      || !SAFE_COLOR_KEY.test(key)
-      || UNSAFE_OBJECT_KEYS.has(key)
-      || !isString(value)
-      || value.length === 0
-      || value.length > MAX_COLOR_VALUE
-      || UNSAFE_CSS.test(value)
-    ) continue
-    out[key] = value
-  }
-  return out
-}
-
-/**
- * Pack any theme records for a v0play hash. Docs, the builder, or a host
- * app can call this — v0play merges `themes` into the sandbox plugin and
- * selects `theme`. Color aliases must already be resolved.
- *
- * @example
- * ```ts
- * toPlaygroundThemes('brand-light', {
- *   'brand-light': { dark: false, colors: { primary: '#7453ec', background: '#ffffff' } },
- *   'brand-dark': { dark: true, colors: { primary: '#c4b5fd', background: '#121212' } },
- * })
- * ```
- */
-export function toPlaygroundThemes (
-  selected: string,
-  records: Record<string, { dark?: boolean, colors?: Record<string, unknown> }>,
-): Pick<PlaygroundHashData, 'theme' | 'themes'> | undefined {
-  const theme = sanitizeThemeId(selected)
-  if (!theme) return undefined
-
-  const themes: Record<string, PlaygroundThemeDefinition> = {}
-  for (const [id, def] of Object.entries(records)) {
-    if (Object.keys(themes).length >= MAX_THEMES) break
-    const safeId = sanitizeThemeId(id)
-    if (!safeId || !def.colors) continue
-    const colors = sanitizeColors(def.colors)
-    if (Object.keys(colors).length === 0) continue
-    themes[safeId] = { dark: def.dark === true, colors }
-  }
-
-  if (Object.keys(themes).length === 0) return { theme }
-  return { theme, themes }
-}
-
-/**
- * Encode editor state (files + active filename) to a URL hash string.
- */
-export async function encodePlaygroundHash (data: PlaygroundHashData): Promise<string> {
-  return utoa(JSON.stringify(data))
-}
-
-function isPlaygroundTheme (v: unknown): v is PlaygroundThemeDefinition {
-  if (!isObject(v)) return false
-  const theme = v as Record<string, unknown>
-  if (!isBoolean(theme.dark) || !isObject(theme.colors)) return false
-  const colors = theme.colors as Record<string, unknown>
-  return Object.values(colors).every(x => isString(x))
-}
-
-function isThemeRecord (v: unknown): v is Record<string, PlaygroundThemeDefinition> {
-  return isObject(v) && Object.values(v as Record<string, unknown>).every(isPlaygroundTheme)
-}
-
-function isValidSettings (v: unknown): v is PlaygroundHashSettings {
-  if (!isObject(v)) return false
-  const settings = v as Record<string, unknown>
-  if (!(isUndefined(settings.vue) || isString(settings.vue))) return false
-  if (!(isUndefined(settings.v0) || isString(settings.v0))) return false
-  if (!(isUndefined(settings.vuetify) || isString(settings.vuetify))) return false
-  if (!(isUndefined(settings.vuetifyNightly) || isBoolean(settings.vuetifyNightly))) return false
-  if (!(isUndefined(settings.preset) || isString(settings.preset))) return false
-  if (!(isUndefined(settings.addons) || isString(settings.addons))) return false
-  return true
-}
-
-function readPlaygroundThemes (parsed: Record<string, unknown>): Pick<PlaygroundHashData, 'theme' | 'themes'> {
-  const settings = isObject(parsed.settings) ? parsed.settings as Record<string, unknown> : undefined
-  const theme = isString(parsed.theme)
-    ? parsed.theme
-    : (isString(settings?.theme) ? settings.theme : undefined)
-  const themesRaw = parsed.themes ?? settings?.themes
-  return {
-    theme,
-    themes: isThemeRecord(themesRaw) ? themesRaw : undefined,
-  }
-}
-
 /**
  * Decode an editor hash back to editor state.
  * Handles 4 formats:
  * 1. Legacy plain Record<string, string>
- * 2 & 3. Current object { files, active, imports, settings? }
+ * 2 & 3. Current object { files, active, imports, settings?, theme?, themes? }
  * 4. Vuetify play tuple [files, vueVersion, vuetifyVersion, appendJson, activeFile, ...]
  */
 export async function decodePlaygroundHash (hash: string): Promise<PlaygroundHashData | null> {
   try {
-    const parsed: unknown = JSON.parse(await atou(hash))
+    const parsed: unknown = await unzipPlaygroundHash(hash)
 
-    // Format 1: legacy plain Record<string, string>
-    if (isFileRecord(parsed)) {
-      return { files: parsed }
-    }
-
-    // Format 4: Vuetify play tuple [files, vueVersion, vuetifyVersion, appendJson, activeFile, ...]
     if (isArray(parsed)) {
       const result = parseVuetifyPlayTuple(parsed)
       if (!result) return null
@@ -406,27 +185,7 @@ export async function decodePlaygroundHash (hash: string): Promise<PlaygroundHas
       return { files: result.files, active: result.active, imports, settings }
     }
 
-    // Formats 2 & 3: current object { files, active, imports, settings? }
-    if (
-      typeof parsed === 'object'
-      && parsed !== null
-      && 'files' in parsed
-      && isFileRecord((parsed as { files: unknown }).files)
-    ) {
-      const record = parsed as Record<string, unknown>
-      const { files, active, imports, settings } = record
-      const packed = readPlaygroundThemes(record)
-      return {
-        files: files as Record<string, string>,
-        active: isString(active) ? active : undefined,
-        imports: isFileRecord(imports) ? imports : undefined,
-        settings: isValidSettings(settings) ? settings : undefined,
-        theme: packed.theme,
-        themes: packed.themes,
-      }
-    }
-
-    return null
+    return parsePlaygroundPayload(parsed)
   } catch {
     return null
   }
