+++ apps/playground/src/composables/usePlaygroundFiles.ts
@@ -1,13 +1,15 @@
+import { isPlaygroundThemeId, sanitizePlaygroundThemes } from '@vuetify/play'
+
 // Framework
 import { IN_BROWSER, isArray, isObject, useTheme, useTimer } from '@vuetify/v0'
 
 // Composables
 import { readPlaygroundIdFromUrl, useOnePlaygrounds, usePlaygroundRouteId } from '@/composables/useOnePlaygrounds'
-import { decodePlaygroundHash, encodePlaygroundHash, isFileRecord, parseVuetifyPlayTuple, SAFE_THEME_ID } from '@/composables/usePlayground'
+import { decodePlaygroundHash, encodePlaygroundHash, isFileRecord, parseVuetifyPlayTuple } from '@/composables/usePlayground'
 import { usePlaygroundSettings } from '@/composables/usePlaygroundSettings'
 
 // Data
-import { createMainTs, createVuetifyTs, REPL_BUILTIN_FILES, REPL_TSCONFIG, REPL_TYPESCRIPT_VERSION, sanitizePlaygroundThemes, UNO_CONFIG_TS, vuetifyEsmUrl } from '@/data/playground-defaults'
+import { createMainTs, createVuetifyTs, REPL_BUILTIN_FILES, REPL_TSCONFIG, REPL_TYPESCRIPT_VERSION, UNO_CONFIG_TS, vuetifyEsmUrl } from '@/data/playground-defaults'
 import { ADDONS, DEFAULT_APP, PRESETS } from '@/data/presets'
 import { parseRegistryQuery, resolveRegistryExample } from '@/data/registry'
 import { parseVuetifyExampleQuery, resolveVuetifyExample } from '@/data/vuetify-examples'
@@ -90,9 +92,7 @@ export function usePlaygroundFiles () {
 
   function applyIncomingTheme (data?: Pick<PlaygroundHashData, 'theme' | 'themes' | 'settings'>) {
     const raw = data?.theme ?? (data?.settings as { theme?: string } | undefined)?.theme
-    const id = raw && SAFE_THEME_ID.test(raw) && raw !== 'constructor' && raw !== 'prototype'
-      ? raw
-      : undefined
+    const id = raw && isPlaygroundThemeId(raw) ? raw : undefined
     extraDefault.value = id
     extraThemes.value = sanitizePlaygroundThemes(
       data?.themes ?? (data?.settings as { themes?: Record<string, PlaygroundThemeDefinition> } | undefined)?.themes,
