+++ apps/playground/src/data/playground-defaults.ts
@@ -1,3 +1,8 @@
+import { isPlaygroundThemeId, sanitizePlaygroundThemes } from '@vuetify/play'
+
+// Types
+import type { PlaygroundThemeDefinition } from '@vuetify/play'
+
 // ── Infrastructure file constants ────────────────────────────────────────
 
 /** Files generated by the editor infrastructure (not user-created) */
@@ -98,55 +103,7 @@ function vuetifyCssUrl (version = 'latest', nightly = false): string {
   return `https://cdn.jsdelivr.net/npm/${pkg}@${version}/dist/vuetify-labs.css`
 }
 
-interface SandboxTheme {
-  dark: boolean
-  colors: Record<string, string>
-}
-
-const SAFE_THEME_ID = /^[a-zA-Z][\w-]*$/
-const SAFE_COLOR_KEY = /^[a-zA-Z0-9_-]+$/
-const UNSAFE_CSS = /url\s*\(|src\s*\(|image\s*\(|image-set\s*\(|cross-fade\s*\(|@import|expression\s*\(|[;{}<>\\]|\/\*/i
-const UNSAFE_OBJECT_KEYS = new Set(['__proto__', 'constructor', 'prototype'])
-const MAX_THEME_ID = 64
-const MAX_THEMES = 32
-const MAX_COLORS = 64
-const MAX_COLOR_VALUE = 128
-
-function isSafeThemeId (id: string): boolean {
-  return id.length <= MAX_THEME_ID && SAFE_THEME_ID.test(id) && !UNSAFE_OBJECT_KEYS.has(id)
-}
-
-function isSafeColorKey (key: string): boolean {
-  return key.length <= MAX_THEME_ID && SAFE_COLOR_KEY.test(key) && !UNSAFE_OBJECT_KEYS.has(key)
-}
-
-function isSafeColorValue (value: string): boolean {
-  return value.length > 0 && value.length <= MAX_COLOR_VALUE && !UNSAFE_CSS.test(value)
-}
-
-/** Strip attacker-controlled theme records down to ids/colors the sandbox will emit. */
-export function sanitizePlaygroundThemes (
-  extra?: Record<string, SandboxTheme>,
-): Record<string, SandboxTheme> | undefined {
-  if (!extra) return undefined
-
-  const out: Record<string, SandboxTheme> = {}
-  for (const [id, def] of Object.entries(extra)) {
-    if (!isSafeThemeId(id) || !def?.colors || Object.keys(out).length >= MAX_THEMES) continue
-    const colors: Record<string, string> = {}
-    for (const [key, value] of Object.entries(def.colors)) {
-      if (!isSafeColorKey(key) || typeof value !== 'string' || !isSafeColorValue(value)) continue
-      if (Object.keys(colors).length >= MAX_COLORS) break
-      colors[key] = value
-    }
-    if (Object.keys(colors).length === 0) continue
-    out[id] = { dark: def.dark === true, colors }
-  }
-
-  return Object.keys(out).length > 0 ? out : undefined
-}
-
-const DEFAULT_SANDBOX_THEMES: { light: SandboxTheme, dark: SandboxTheme } = {
+const DEFAULT_SANDBOX_THEMES: { light: PlaygroundThemeDefinition, dark: PlaygroundThemeDefinition } = {
   light: {
     dark: false,
     colors: {
@@ -204,8 +161,8 @@ const DEFAULT_SANDBOX_THEMES: { light: SandboxTheme, dark: SandboxTheme } = {
 }
 
 function mergeSandboxThemes (
-  extra?: Record<string, SandboxTheme>,
-): Record<string, SandboxTheme> {
+  extra?: Record<string, PlaygroundThemeDefinition>,
+): Record<string, PlaygroundThemeDefinition> {
   return {
     light: { dark: false, colors: { ...DEFAULT_SANDBOX_THEMES.light.colors } },
     dark: { dark: true, colors: { ...DEFAULT_SANDBOX_THEMES.dark.colors } },
@@ -213,7 +170,7 @@ function mergeSandboxThemes (
   }
 }
 
-export function createMainTs (defaultTheme = 'light', options?: MainOptions, vuetifyVersion = 'latest', vuetifyNightly = false, extraThemes?: Record<string, SandboxTheme>): string {
+export function createMainTs (defaultTheme = 'light', options?: MainOptions, vuetifyVersion = 'latest', vuetifyNightly = false, extraThemes?: Record<string, PlaygroundThemeDefinition>): string {
   const useV0 = options?.v0 !== false
   const extraImports: string[] = []
   const extraPlugins: string[] = []
@@ -261,7 +218,7 @@ export function createMainTs (defaultTheme = 'light', options?: MainOptions, vue
   const setupBlock = extraSetup.length > 0 ? '\n' + extraSetup.join('\n') + '\n' : ''
   const pluginBlock = extraPlugins.length > 0 ? extraPlugins.join('\n') + '\n' : ''
   const mergedThemes = mergeSandboxThemes(extraThemes)
-  const safeDefault = isSafeThemeId(defaultTheme) && Object.hasOwn(mergedThemes, defaultTheme)
+  const safeDefault = isPlaygroundThemeId(defaultTheme) && Object.hasOwn(mergedThemes, defaultTheme)
     ? defaultTheme
     : 'light'
   const themesJson = JSON.stringify(mergedThemes, null, 2).split('\n').join('\n  ')
