+++ packages/0/src/composables/useTheme/index.test.ts
@@ -933,6 +933,18 @@ describe('createTheme', () => {
 
       expect(css).toContain('color-scheme: dark')
     })
+
+    it('should omit color values that contain css comments', () => {
+      const adapter = new V0StyleSheetThemeAdapter()
+
+      const css = adapter.generate(
+        { light: { primary: '#1976d2', inject: 'red /*' } },
+        false,
+      )
+
+      expect(css).toContain('--v0-primary: #1976d2')
+      expect(css).not.toContain('red /*')
+    })
   })
 
   describe('register with colors', () => {
