+++ packages/play/src/files.ts
@@ -1,0 +1,75 @@
+export interface PlaygroundFile {
+  name: string
+  code: string
+}
+
+function toPascal (str: string): string {
+  return str.replace(/(^|-)(\w)/g, (_, __, c: string) => c.toUpperCase())
+}
+
+/**
+ * Detect which file is the entry point for a multi-file example.
+ */
+export function detectEntryFile (files: PlaygroundFile[]): PlaygroundFile | undefined {
+  const vueFiles = files.filter(f => f.name.endsWith('.vue'))
+
+  const entryNames = ['index.vue', 'App.vue', 'example.vue', 'main.vue']
+  for (const name of entryNames) {
+    const found = vueFiles.find(f => f.name.toLowerCase() === name.toLowerCase())
+    if (found) return found
+  }
+
+  for (const file of vueFiles) {
+    const importsOthers = vueFiles.some(other =>
+      other !== file && file.code.includes(`./${other.name.replace(/\.\w+$/, '')}`),
+    )
+    if (importsOthers) return file
+  }
+
+  return vueFiles.at(-1)
+}
+
+/**
+ * Generate an App.vue wrapper that imports and renders the entry component.
+ */
+export function generateAppWrapper (entryPath: string): string {
+  const baseName = entryPath.split('/').pop()!.replace(/\.vue$/, '')
+  const pascalName = toPascal(baseName)
+  return [
+    '<' + `script setup lang="ts">`,
+    `  import ${pascalName} from './${entryPath}'`,
+    '</' + 'script>',
+    '',
+    '<template>',
+    '  <div class="p-4">',
+    `    <${pascalName} />`,
+    '  </div>',
+    '</template>',
+    '',
+  ].join('\n')
+}
+
+/**
+ * Build the src/-prefixed file record that the playground REPL expects.
+ * When dir is provided, files are nested: src/{dir}/{name}
+ */
+export function buildPlaygroundFiles (inputFiles: PlaygroundFile[], dir?: string): Record<string, string> {
+  const files: Record<string, string> = {}
+  const prefix = dir ? `src/${dir}` : 'src'
+
+  for (const file of inputFiles) {
+    const path = file.name.startsWith('src/') ? file.name : `${prefix}/${file.name}`
+    files[path] = file.code
+  }
+
+  const hasAppVue = inputFiles.some(f => f.name.toLowerCase() === 'app.vue')
+  if (!hasAppVue) {
+    const entryFile = detectEntryFile(inputFiles)
+    if (entryFile) {
+      const entryPath = dir ? `${dir}/${entryFile.name}` : entryFile.name
+      files['src/App.vue'] = generateAppWrapper(entryPath)
+    }
+  }
+
+  return files
+}
