+++ packages/play/src/vite-env.d.ts
@@ -1,0 +1,3 @@
+declare const __DEV__: boolean
+declare const __VERSION__: string
+declare const __VITE_LOGGER_ENABLED__: string
