+++ packages/play/tsconfig.json
@@ -1,16 +1,15 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
   "compilerOptions": {
+    "noEmit": false,
+    "allowImportingTsExtensions": false,
     "composite": true,
     "incremental": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "customConditions": ["development"],
-    "moduleResolution": "bundler",
     "paths": {
-      "@/*": ["./src/*"],
-      "@vuetify/play": ["../../packages/play/src/index.ts"],
-      "#v0": ["../../packages/0/src"],
-      "#v0/*": ["../../packages/0/src/*"]
+      "#v0": ["../0/src"],
+      "#v0/*": ["../0/src/*"]
     },
     "lib": ["dom", "dom.iterable", "esnext"],
     "skipLibCheck": true,
@@ -21,19 +20,11 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
+    "moduleResolution": "bundler",
     "module": "esnext",
-    "target": "esnext"
+    "target": "esnext",
+    "allowSyntheticDefaultImports": true
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.vue"
-  ],
-  "exclude": [
-    "src/**/__tests__/*",
-    "src/util/export-templates/uno.config.ts",
-    "src/util/export-templates/vite.config.ts"
-  ],
-  "references": [
-    { "path": "../../packages/0" }
-  ]
+  "include": ["src/**/*.ts"],
+  "exclude": ["src/**/*.test.ts"]
 }
